refactor:update serve for pro
